rename get serv status msg
